+++ backend/src/main/java/com/ssafy/hellojob/domain/jobroleanalysis/controller/JobRoleAnalysisController.java
@@ -1,21 +1,19 @@
 package com.ssafy.hellojob.domain.jobroleanalysis.controller;
 
-import com.ssafy.hellojob.domain.jobroleanalysis.dto.*;
-import com.ssafy.hellojob.domain.jobroleanalysis.entity.JobRoleAnalysisBookmark;
+import com.ssafy.hellojob.domain.jobroleanalysis.dto.request.JobRoleAnalysisBookmarkSaveRequestDto;
+import com.ssafy.hellojob.domain.jobroleanalysis.dto.request.JobRoleAnalysisSaveRequestDto;
+import com.ssafy.hellojob.domain.jobroleanalysis.dto.request.JobRoleAnalysisSearchCondition;
+import com.ssafy.hellojob.domain.jobroleanalysis.dto.request.JobRoleAnalysisUpdateRequestDto;
+import com.ssafy.hellojob.domain.jobroleanalysis.dto.response.*;
 import com.ssafy.hellojob.domain.jobroleanalysis.service.JobRoleAnalysisService;
-import com.ssafy.hellojob.domain.user.entity.User;
-import com.ssafy.hellojob.domain.user.service.UserService;
 import com.ssafy.hellojob.global.auth.token.UserPrincipal;
 import com.ssafy.hellojob.global.exception.BaseException;
 import com.ssafy.hellojob.global.exception.ErrorCode;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
-import java.security.Principal;
 import java.util.List;
 
 @RestController
@@ -43,7 +41,7 @@ public class JobRoleAnalysisController {
 
     // 직무 분석 상세 조회
     @GetMapping("/{jobRoleAnalysisId}")
-    public JobRoleAnalysisDetailResponseDto jobRoleAnalysisDetail(@PathVariable("jobRoleAnalysisId") Long jobRoleAnalysisId,
+    public JobRoleAnalysisDetailResponseDto jobRoleAnalysisDetail(@PathVariable("jobRoleAnalysisId") Integer jobRoleAnalysisId,
                                                                   @AuthenticationPrincipal UserPrincipal userPrincipal){
 
         Integer userId = userPrincipal.getUserId();
@@ -64,7 +62,7 @@ public class JobRoleAnalysisController {
 
     // 직무 분석 북마크 삭제
     @DeleteMapping("/bookmark/{jobRoleAnalysisBookmarkId}")
-    public void JobRoleAnalysisBookmarkDelete (@PathVariable("jobRoleAnalysisBookmarkId") Long jobRoleAnalysisBookmarkId,
+    public void JobRoleAnalysisBookmarkDelete (@PathVariable("jobRoleAnalysisBookmarkId") Integer jobRoleAnalysisBookmarkId,
                                                @AuthenticationPrincipal UserPrincipal userPrincipal){
         Integer userId = userPrincipal.getUserId();
 
@@ -73,8 +71,8 @@ public class JobRoleAnalysisController {
 
     // 직무 분석 북마크 목록 조회
     @GetMapping("/bookmark")
-    public List<JobRoleAnalysisListResponseDto> JobRoleAnalysisBookmarkList(@RequestParam(value = "companyId", required = false) Long companyId,
-                                                         @AuthenticationPrincipal UserPrincipal userPrincipal) {
+    public List<JobRoleAnalysisListResponseDto> JobRoleAnalysisBookmarkList(@RequestParam(value = "companyId", required = false) Integer companyId,
+                                                                            @AuthenticationPrincipal UserPrincipal userPrincipal) {
         Integer userId = userPrincipal.getUserId();
         List<JobRoleAnalysisListResponseDto> result;
 
@@ -90,9 +88,9 @@ public class JobRoleAnalysisController {
 
     // 직무 분석 검색(기본값: companyId, 검색 조건: 직무명, 직무 분석 제목, 직무 카테고리)
     @GetMapping("/{companyId}/search")
-    public List<JobRoleAnalysisSearchListResponseDto> JobRoleAnalysisSearch(@PathVariable Long companyId,
-                                                   @ModelAttribute JobRoleAnalysisSearchCondition condition,
-                                                   @AuthenticationPrincipal UserPrincipal userPrincipal) {
+    public List<JobRoleAnalysisSearchListResponseDto> JobRoleAnalysisSearch(@PathVariable Integer companyId,
+                                                                            @ModelAttribute JobRoleAnalysisSearchCondition condition,
+                                                                            @AuthenticationPrincipal UserPrincipal userPrincipal) {
 
         Integer userId = userPrincipal.getUserId();
         List<JobRoleAnalysisSearchListResponseDto> result = jobRoleAnalysisService.searchJobRoleAnalysis(userId, companyId, condition);
@@ -102,7 +100,7 @@ public class JobRoleAnalysisController {
 
     // 직무 분석 삭제
     @DeleteMapping("/analysis/{jobRoleAnalysisId}")
-    public void JobRoleAnalysisDelete(@PathVariable Long jobRoleAnalysisId,
+    public void JobRoleAnalysisDelete(@PathVariable Integer jobRoleAnalysisId,
                                       @AuthenticationPrincipal UserPrincipal userPrincipal){
         Integer userId = userPrincipal.getUserId();
         jobRoleAnalysisService.deleteJobRoleAnalysis(userId, jobRoleAnalysisId);
